Docker images including DGL itself (#36)
This contains an image that just has the dependencies necessary for
building and running DGL on ROCm for use cases like mounting a local git
repository as well as an image that builds and installs DGL itself for a
more turnkey experience.

I played with various ways to do this. The pytorch/rocm image already
has its own conda env that it has sort of activated in a djanky way, so
we use that instead of attempting to create our own (we also need the
version of torch it has). Using the DGL script for creating and
activating the conda env was silently failing because it tried to
install conda when it was already installed.

This is configured to build with all the AMD GPU architectures we can,
which unfortunately isn't very many, as we are constrained by hip
collections. Even with just 4 architectures it's quite slow to build DGL
though, especially graphbolt. We can build and provide this general
image, but I included instructions for users to build with their own set
of architectures.