Cache apt and pip in docker builds (#41)
I also looked at multistage builds, but in the end wasn't able to get
anything sufficiently useful out of them to justify the added
complexity. I think if we controlled the original base image we might be
able to, but as it is, everything big ends up scattered all over the
filesystem.

The good news is that the caches do speed things up quite a bit, even if
they don't help with image size.