Adding dependency info to wheel names for rocm builds
We need to add a dependency info to the wheen file names for rocm
builds to differentiate them. For now, we're just adding them with the
pre-existing env variable DGL_PACKAGE_SUFFIX. This makes the wheel file
names rather unweildy, but until we start using something like the wheel
variants and manylinux it's what we're stuck with.