Updating version (2.4.0.amd0) and version scripts
We've updated how we wanted to do AMD versioning. The current plan is to
do <upstream_version>.amd<amd_release_number>.

We've also updated the utils to update the version to handle the
compilcated DGL_VERSION_SUFFIX env variables we pass. We want the final
python wheels to be PEP440 compliant so we have to transform the version
to <upstream_version>+amd<amd_release_number>.