Bumping hipcollections to the latest version
This commit (sloppily) adds support for the latest hipcollections
release (release/rocmds-25.10). There are all sorts of potholes here,
hipcollections + libhipcxx + rocm need special care to work correctly.

I've opened issued in libhipcxx to try and address some of these but for
now, we handle it with some source file manipulation in our script to
install graphbolt dependencies.

This PR will break backward compatibility with rocm < 7, but that should
be ok because we'll probably be focused on rocm@7.9+ in the next
release.